Use a separate browser session for each page
The official site stores a lot of state in the session. This can lead to
situations where the session becomes invalid and the site starts
returning pages with no content.

This change fixes this problem by using a new browser instance for each
request, which sidesteps any problems with changing session variables.